+++ examples/configs/src/view-configs/empty.js
@@ -1,0 +1,14 @@
+export const emptyConfig = {
+  name: 'Drag and drop files to get started',
+  description: '',
+  version: '1.0.17',
+  initStrategy: 'auto',
+  datasets: [
+
+  ],
+  coordinationSpace: {
+
+  },
+  layout: [
+  ],
+};