+++ packages/config/src/generate-config.js
@@ -1,12 +1,18 @@
 // TODO: ts-check
+
 import { FileType } from '@vitessce/constants-internal';
-import { withConsolidated, FetchStore, ZipFileStore, open as zarrOpen, root as zarrRoot } from 'zarrita';
+import { withConsolidated, FetchStore, open as zarrOpen, root as zarrRoot } from 'zarrita';
+// eslint-disable-next-line import/no-unresolved
+import ZipFileStore from '@zarrita/storage/zip';
 import { transformEntriesForZipFileStore } from '@vitessce/zarr-utils';
 import { VitessceConfig } from './VitessceConfig.js';
 // Classes for different types of objects
 import { AnnDataAutoConfig } from './generate-config-anndata.js';
 import { SpatialDataAutoConfig } from './generate-config-spatialdata.js';
 import { OmeAutoConfig } from './generate-config-ome.js';
+
+// TODO: make this a function parameter?
+const FILE_TYPE_DELIM = '$';
 
 const fileTypeToExtensions = {
   [FileType.IMAGE_OME_TIFF]: ['.ome.tif', '.ome.tiff', '.ome.tf2', '.ome.tf8'],
@@ -51,6 +57,9 @@ const ZARR_FILETYPES = [
 ];
 
 function urlToFileType(url) {
+  // TODO: for plain .zarr, we could open the root .attrs and try to guess a fileType
+  // (we can probably infer at least whether OME-Zarr vs. AnnData vs. SpatialData).
+
   const match = Object.entries(fileTypeToExtensions).find(
     // eslint-disable-next-line no-unused-vars
     ([fileType, extensions]) => extensions.some(ext => url.endsWith(ext)),
@@ -99,15 +108,12 @@ function ensureStores(parsedUrls) {
 
 /**
  *
- * @param {string} s A single string, like this
- * `http://example.com/my_zarr.zarr#anndata.zarr;
- * http://example.com/my_tiff.ome.tif`
- * @returns {{ url: string, fileType: string}[]} The URLs with file types.
+ * @param {string[]} arr
+ * @returns {{ url: string, fileType: string }[]} The URLs with file types.
  */
-export function parseUrls(s) {
-  const urlsWithHashes = s.split(';');
-  return urlsWithHashes.map((urlWithHash) => {
-    const parts = urlWithHash.split('#');
+export function parseUrls(arr) {
+  return arr.map((urlWithHash) => {
+    const parts = urlWithHash.split(FILE_TYPE_DELIM);
     if (parts.length === 1) {
       const [url] = parts;
       return {
@@ -121,8 +127,20 @@ export function parseUrls(s) {
         fileType,
       };
     }
-    throw new Error('Only expected zero or one # character per URL, but received more.');
+    throw new Error(`Only expected zero or one ${FILE_TYPE_DELIM} character per URL, but received more.`);
   });
+}
+
+/**
+ *
+ * @param {string} s A single string, like this
+ * `http://example.com/my_zarr.zarr#anndata.zarr;
+ * http://example.com/my_tiff.ome.tif`
+ * @returns {{ url: string, fileType: string}[]} The URLs with file types.
+ */
+export function parseUrlsFromString(s) {
+  const urlsWithHashes = s.split(';');
+  return parseUrls(urlsWithHashes);
 }
 
 
@@ -247,7 +265,7 @@ export async function generateConfig(parsedUrls, layoutOption = null) {
 
     autoConfig.addFiles(vc, dataset);
     // TODO: add all files, then add all views (in two separate loops)?
-    autoConfig.addViews(vc, layoutOption);
+    autoConfig.addViews(vc, dataset, layoutOption);
   });
 
   const stores = Object.fromEntries(